+++ GloryEngine/Editor/GloryEditor/SceneObjectPropertyDrawer.h
@@ -17,7 +17,7 @@ namespace Glory::Editor
         SceneObjectRedirectPropertyDrawer() : PropertyDrawer(ST_Object) {}
 
     protected:
-        bool Draw(const std::string& label, void* data, size_t typeHash, uint32_t flags) const override;
-        bool Draw(const std::string& label, YAML::Node& node, size_t typeHash, uint32_t flags) const override;
+        bool Draw(const std::string& label, void* data, uint32_t typeHash, uint32_t flags) const override;
+        bool Draw(const std::string& label, YAML::Node& node, uint32_t typeHash, uint32_t flags) const override;
     };
 }
