+++ GloryEngine/Engine/GloryCore/LayerRef.cpp
@@ -25,7 +25,7 @@ namespace Glory
 	const GloryReflect::TypeData* LayerRef::GetTypeData()
 	{
 		static const char* typeNameString = "LayerRef";
-		static const size_t TYPE_HASH = ResourceType::GetHash<LayerRef>();
+		static const uint32_t TYPE_HASH = ResourceType::GetHash<LayerRef>();
 		static const int NUM_ARGS = 0;
 		static const GloryReflect::TypeData pTypeData = GloryReflect::TypeData(typeNameString, nullptr, TYPE_HASH, NUM_ARGS);
 		return &pTypeData;
